src/main/java/dev/jbriggs/aoc/handheld/reader/MarkerReader.java
package dev.jbriggs.aoc.handheld.reader;

import dev.jbriggs.aoc.handheld.DeviceException;
import java.util.List;
import lombok.Getter;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@Getter
public class MarkerReader implements Reader {

  private String input;

  @Override
  public void readAll(List<String> input) throws ReaderException {
    this.input = input.get(0);
  }

  public int findMarkerPosition(int markerLength) {
    char[] chars = input.toCharArray();
    for (int i = markerLength; i < chars.length; i++) {
      if (hasUniqueValues(chars, i, markerLength)) {
        return i;
      }
    }
    return input.length();
  }

  private static boolean hasUniqueValues(char[] data,
      int currentCharacterPosition, int markerLength) {
    for (int toCheck = currentCharacterPosition - (markerLength);
        toCheck < currentCharacterPosition; toCheck++) {
      for (int other = toCheck + 1; other < currentCharacterPosition; other++) {
        char otherCharacter = data[other];
        char toCheckCharacter = data[toCheck];
        if (toCheckCharacter == otherCharacter && other != toCheck) {
          return false;
        }
      }
    }
    return true;
  }
}
